requirements.txt
PyQt5==5.15.9
hijri-converter==2.3.1
python-dotenv==1.0.0
reportlab==4.0.4
twilio==8.5.0
Pillow==10.0.0
arabic-reshaper==3.0.0
python-bidi==0.4.2
cryptography==41.0.1
pandas==2.0.3
openpyxl==3.1.2
qrcode==7.4.2